174-FE: Fix testing errors caused by circular import in `FINANCES_COPY`
