+++ Okane.Client/src/features/financeRecords/components/searchFilters/AppliedDescriptionFilter.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 // Internal
-import { FINANCES_COPY } from '@features/financeRecords/constants/copy'
+import { FINANCES_SEARCH_FILTERS_COPY } from '@features/financeRecords/constants/searchFiltersCopy'
 
 import { type FinanceRecordSearchFilters } from '@features/financeRecords/types/searchFilters'
 
@@ -11,6 +11,6 @@ const props = defineProps<AppliedDescriptionFilterProps>()
 
 <template>
   <li v-if="props.description">
-    {{ FINANCES_COPY.SEARCH_FILTERS.APPLIED_DESCRIPTION({ description: props.description }) }}
+    {{ FINANCES_SEARCH_FILTERS_COPY.APPLIED_DESCRIPTION({ description: props.description }) }}
   </li>
 </template>
