+++ Okane.Client/src/features/financeRecords/components/searchFilters/AppliedHappenedAtFilter.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 // Internal
-import { FINANCES_COPY } from '@features/financeRecords/constants/copy'
+import { FINANCES_SEARCH_FILTERS_COPY } from '@features/financeRecords/constants/searchFiltersCopy'
 
 import { type FinanceRecordSearchFilters } from '@features/financeRecords/types/searchFilters'
 
@@ -15,7 +15,7 @@ const props = defineProps<AppliedHappenedAtFilterProps>()
 <template>
   <li v-if="props.happenedAtOperator && props.happenedAt1">
     {{
-      FINANCES_COPY.SEARCH_FILTERS.APPLIED_HAPPENED_AT_AND_OPERATOR({
+      FINANCES_SEARCH_FILTERS_COPY.APPLIED_HAPPENED_AT_AND_OPERATOR({
         happenedAt: props.happenedAt1,
         operator: props.happenedAtOperator,
       })
@@ -23,7 +23,7 @@ const props = defineProps<AppliedHappenedAtFilterProps>()
   </li>
   <li v-else-if="props.happenedAt1 && props.happenedAt2">
     {{
-      FINANCES_COPY.SEARCH_FILTERS.APPLIED_HAPPENED_AT_RANGE({
+      FINANCES_SEARCH_FILTERS_COPY.APPLIED_HAPPENED_AT_RANGE({
         happenedAt1: props.happenedAt1,
         happenedAt2: props.happenedAt2,
       })
