+++ Okane.Client/src/features/financeRecords/components/searchFilters/AppliedSorting.spec.ts
@@ -3,7 +3,7 @@ import AppliedSorting, {
   type AppliedSortingProps,
 } from '@features/financeRecords/components/searchFilters/AppliedSorting.vue'
 
-import { FINANCES_COPY } from '@features/financeRecords/constants/copy'
+import { FINANCES_SEARCH_FILTERS_COPY } from '@features/financeRecords/constants/searchFiltersCopy'
 import { SORT_DIRECTION } from '@shared/constants/search'
 
 const mountComponent = getMountComponent(AppliedSorting)
@@ -16,7 +16,7 @@ test('renders the sort field and direction', () => {
   const wrapper = mountComponent({ props })
   const lis = wrapper.findAllByText(
     'li',
-    FINANCES_COPY.SEARCH_FILTERS.APPLIED_SORTING({
+    FINANCES_SEARCH_FILTERS_COPY.APPLIED_SORTING({
       sortDirection: props.sortDirection,
       sortField: props.sortField,
     }),
