+++ Okane.Client/src/features/financeRecords/components/searchFilters/AppliedSorting.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 // Internal
-import { FINANCES_COPY } from '@features/financeRecords/constants/copy'
+import { FINANCES_SEARCH_FILTERS_COPY } from '@features/financeRecords/constants/searchFiltersCopy'
 
 import { type FinanceRecordSearchFilters } from '@features/financeRecords/types/searchFilters'
 
@@ -12,7 +12,7 @@ const props = defineProps<AppliedSortingProps>()
 <template>
   <li>
     {{
-      FINANCES_COPY.SEARCH_FILTERS.APPLIED_SORTING({
+      FINANCES_SEARCH_FILTERS_COPY.APPLIED_SORTING({
         sortDirection: props.sortDirection,
         sortField: props.sortField,
       })
